+++ CHANGELOG.md
@@ -113,6 +113,9 @@ CHANGELOG
 ### Tools
 
 ### Contrib
+- Check MK Statistics Cronjob:
+  - Use `statistics_*` parameters.
+  - Make file executable
 
 ### Known issues